fix: precompiled headers, animation shader
Added precompiled headers with C++ headers and other libraries that will not be modified. Animation frames now get properly displayed: issue 1. hard-coded values inside CPU code for UV's and issue 2. weird offsets inside the GPU code for the UV's. Apart from those, general clean-up for unused files files.

+++ Headers/QuadArray.hpp
@@ -10,7 +10,7 @@
 class QuadArray
 {
 public:
-    QuadArray(uint8_t MaxNumQuads);
+    QuadArray(int MaxNumQuads);
 
     ~QuadArray();
