+++ Headers/Texture2D.hpp
@@ -1,14 +1,13 @@
 #pragma once
 #include <glad/glad.h>
-#include <cstdint>
 
 struct SpriteInfo
 {
-    uint8_t PixelX = 0;
-    uint8_t PixelY = 0;
-    uint8_t SpriteRow = 0;    // Sprite location in the sprite sheet - row
-    uint8_t SpriteCol = 0;    // Sprite location in the sprite sheet - col
-    uint8_t SpriteWidth = 0;
+    int PixelX = 0;
+    int PixelY = 0;
+    int SpriteRow = 0;    // Sprite location in the sprite sheet - row
+    int SpriteCol = 0;    // Sprite location in the sprite sheet - col
+    int SpriteWidth = 0;
 };
 
 class Texture2D
